+++ CHANGELOG.md
@@ -28,6 +28,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
   - You no longer can use `nil` values on a property with `Grax.put_additional_statements/2`
     to remove statements with this property. You must use the new 
     `Grax.delete_additional_predicates/2` function for this now.
+- "heterogeneous link properties" are now called "polymorphic link properties"
+  (since this name didn't appear in the code, this change only affects the documentation)
 
 ### Fixed